requirements.txt
requests
openai-agents
python-dotenv
galileo[openai]

